pyproject.toml
[project]
name = "gcode-security-guard"
version = "0.1.0"
description = "Gcode Security Guard: 麒麟OS智能运维Agent — 意图过滤 + 审计系统"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "pyyaml>=6.0",
    "transformers>=4.46",
    "torch>=2.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "ruff>=0.8",
]

[build-system]
requires = ["setuptools>=64.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
